backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#

django==2.2.6
attrs==19.3.0             # via pytest
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
#django==2.2.6
idna==2.8                 # via requests
importlib-metadata==1.5.0  # via pluggy, pytest
more-itertools==8.2.0     # via pytest
packaging==20.1           # via pytest
pillow==7.0.0
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pyparsing==2.4.6          # via packaging
pytest
pyjwt==2.1.0
pillow==7.0.0
pytest-pythonpath==0.7.3
pytest-django
django-filter==2.4.0
asgiref==3.2.10
python-dotenv==0.19.2
sorl-thumbnail==12.6.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
django-colorfield==0.4.5
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
requests==2.22.0
six==1.14.0               # via packaging
drf-extra-fields==3.1.1
django-wkhtmltopdf
urllib3==1.25.6           # via requests
wcwidth==0.1.8            # via pytest
zipp==2.2.0               # via importlib-metadata
mixer==7.1.2
django-environ==0.8.1
djoser==2.1.0
